Added a new resource
Hi, I have added a new resource on C Structures. I found this tutorial on structures in C online while looking for some resources to learn C programming. Topics were explained in detail and found it to be useful and thought of adding it here. Hope this small contribution will help other learners.